Ver 0.01 - Analog Input and Output Tested and Working
All the basics work with NIMEX.

1) Most of the old code has been abandoned and the pulser_startPulser
does almost all of the work of the 'controller.'  Things will change
soon, but for now the entire 'session' can be generated by changing a
few parameters in the _startPulser file and running it.

2) Fixed a bad bug in the pulse train and ramp functions.

3) Pulse and ramp functions now return complete vectors bounded by the
desired acquisition time. Before I had the baseline appended to the
train so the user would be free to shape the rest of the train as they
saw fit.  This way of doing things will simplify the UI and code going
forward.

---- Things to add:
Digital I/O
Custom waveform loader
Pulse Train Randomizer
etc.