Ver 0.1 - Synchronized Analog & Digital I/O
Use:
Configure parameters in pulser_startPulser and then run.
Trigger is routed through PFI0 (as usual).

Works on 32 or 64 bit Matlab, but requires DAQmx 4.3 (this is a good
thing because the more recent versions of DAQmx are flaky).

Next Steps:
Could/Should be set-up to route an internal trigger to PFI0
programatically.

Still needs a GUI — ughhh

Will separate the configuration from the program.

Will add more than pulses and ramps.