Ver 0.2 - Multiple Counters/Multiple Cycles
Added suport for a second counter (optional) that can drive a camera at
a rate independent of (but still in sync with) the sample clock.

Added support for multiple trials. Each has to be triggered seperatley.

Major reorg of the code.

Still procrastinating on writing the GUI/Callbacks.

The configuration is still in the main program, not sure if there will
be a better way to do it.  Maybe I can have it in a +… folder and
import it when the function is called … hmm.
The problem lies in wanting to maintain as much of a Model View
Controller scheme as I can, and I don't want to make a hasty choice
I'll regret after making the view(s).